requirements.txt
# Core dependencies
openai>=1.12.0
pymupdf>=1.23.0
requests>=2.31.0
python-dotenv>=1.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Vector database options
faiss-cpu>=1.7.4
chromadb>=0.4.0

# Optional: Supabase integration
supabase>=2.0.0

# Data processing
pandas>=2.0.0
tqdm>=4.65.0

# Testing
pytest>=7.4.0
